CSharp OOP Retake Exam - 18 April 2022/Heroes/Skeleton/Heroes/Repositories/HeroRepository.cs
using Heroes.Models.Contracts;
using Heroes.Repositories.Contracts;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Xml.Linq;

namespace Heroes.Repositories
{
    public class HeroRepository : IRepository<IHero>
    {
        private readonly List<IHero> heroes;

        public HeroRepository()
        {
            this.heroes = new List<IHero>();
        }

        public IReadOnlyCollection<IHero> Models => this.heroes.AsReadOnly();

        public void Add(IHero model)
        {
            this.heroes.Add(model);
        }

        public IHero FindByName(string name)
        {
            return this.heroes.FirstOrDefault(h => h.Name == name);
        }

        public bool Remove(IHero model)
        {
            IHero hero = this.heroes.FirstOrDefault(h => h.Name == model.Name);

            if (hero != null)
            {
                this.heroes.Remove(hero);

                return true;
            }

            return false;
        }
    }
}
